switch to pedro 3
